% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shifts.R
\name{entropy_shift}
\alias{entropy_shift}
\title{Entropy Shift}
\usage{
entropy_shift(
  type2freq_1,
  type2freq_2,
  base = 2L,
  alpha = 1,
  reference_value = 0,
  normalization = "variation"
)
}
\arguments{
\item{type2freq_1}{A data.frame containing words and their frequencies.}

\item{type2freq_2}{A data.frame containing words and their frequencies.}

\item{base}{The base for the logarithm when computing entropy scores.}

\item{alpha}{The parameter for the generalized Tsallis entropy. Setting 'alpha = 1'
recovers the Shannon entropy.}

\item{reference_value}{Optional. String or numeric. The reference score to use
to partition scores into two different regimes. If 'average', uses the average
score according to type2freq_1 and type2score_1. If None and a lexicon is
selected for type2score, uses the respective middle point in that lexicon's scale.
Otherwise if NULL, uses zero as the reference point.}

\item{normalization}{Optional. Default value: "variation". If 'variation', normalizes
shift scores so that the sum of their absolute values sums to 1. If 'trajectory',
normalizes them so that the sum of shift scores is 1 or -1. The trajectory
normalization cannot be applied if the total shift score is 0, so scores are
left unnormalized if the total is 0 and 'trajectory' is specified.}
}
\value{
Returns a list object of class shift.
}
\description{
Shift object for calculating the shift in entropy between two systems.
}
\examples{
example to follow
}
\seealso{
Other shifts: 
\code{\link{jsdivergence_shift}()},
\code{\link{kldivergence_shift}()},
\code{\link{proportion_shift}()},
\code{\link{weighted_avg_shift}()}
}
\concept{shifts}
