% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_bar_colours}
\alias{get_bar_colours}
\title{Get bar colours for plotting}
\usage{
get_bar_colours(top_shift_scores, all_pos_contributions)
}
\arguments{
\item{top_shift_scores}{top n values of the shift_scores data.frame from a
shift object}

\item{all_pos_contributions}{Defaults to FALSE when calling get_shift_graphs.}
}
\value{
Returns a list with 3 values. The values are lists of the colors
to assign to the p_diff and s_diff components respectively. If just the
overall contributions are being shown in a simple (not detailed) shift graph,
then the "total" colours are used.
}
\description{
Returns the component colours of each type's contribution bars.
}
\keyword{internal}
