% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/entropy.R
\name{get_entropy_scores}
\alias{get_entropy_scores}
\title{Get Entropy Scores}
\usage{
get_entropy_scores(type2score_1, type2score_2, base = 2L, alpha = 1)
}
\arguments{
\item{type2score_1}{}

\item{type2score_2}{}

\item{base}{The base for the logarithm when computing entropy scores.}

\item{alpha}{The parameter for the generalized Tsallis entropy. Setting 'alpha = 1'
recovers the Shannon entropy.}
}
\value{
A data.frame containing the entropy scores
}
\description{
Get Entropy Scores
}
\keyword{internal}
