% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/entropy.R
\name{get_entropy_word_scores}
\alias{get_entropy_word_scores}
\title{Get Entropy Word Scores}
\usage{
get_entropy_word_scores(score, base = 2, alpha = 1)
}
\arguments{
\item{score}{The relative frequency.}

\item{base}{The base for the logarithm when computing entropy scores.}

\item{alpha}{The parameter for the generalized Tsallis entropy. Setting 'alpha = 1'
recovers the Shannon entropy.}
}
\value{
The entropy score.
}
\description{
Calculates the generalized Tsallis entropy scores.
}
\keyword{internal}
