% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/entropy.R
\name{get_relative_frequency}
\alias{get_relative_frequency}
\title{Get Relative Frequency}
\usage{
get_relative_frequency(data)
}
\arguments{
\item{data}{A data.frame containing words and their counts}
}
\value{
A data.frame containing words and their relative frequencies
}
\description{
Calculates the relative frequency (proportion) of each word in the data.frame
}
\keyword{internal}
