% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shifts.R
\name{kldivergence_shift}
\alias{kldivergence_shift}
\title{Kullback-Leibler Divergence Shift}
\usage{
kldivergence_shift(
  type2freq_1,
  type2freq_2,
  base = 2L,
  reference_value = 0,
  normalization = "variation"
)
}
\arguments{
\item{type2freq_1}{A data.frame containing words and their frequencies.}

\item{type2freq_2}{A data.frame containing words and their frequencies.}

\item{base}{The base for the logarithm when computing entropy scores.}

\item{reference_value}{Optional. String or numeric. The reference score to use
to partition scores into two different regimes. If 'average', uses the average
score according to type2freq_1 and type2score_1. If a lexicon is used for type2score,
you need to use the middle point of that lexicon's scale. If no value is supplied,
zero will be used as the reference point. See details for more information.}

\item{normalization}{Optional. Default value: "variation". If 'variation', normalizes
shift scores so that the sum of their absolute values sums to 1. If 'trajectory',
normalizes them so that the sum of shift scores is 1 or -1. The trajectory
normalization cannot be applied if the total shift score is 0, so scores are
left unnormalized if the total is 0 and 'trajectory' is specified.}
}
\value{
Returns a list object of class shift.
}
\description{
Shift object for calculating the Kullback-Leibler divergence (KLD) between two systems
}
\examples{
library(shifterator)
library(quanteda)
library(quanteda.textstats)
library(dplyr)

reagan <- corpus_subset(data_corpus_inaugural, President == "Reagan") \%>\% 
  tokens(remove_punct = TRUE) \%>\% 
dfm() \%>\% 
textstat_frequency() \%>\% 
as.data.frame() \%>\% # to move from classes frequency, textstat, and data.frame to data.frame
select(feature, frequency) 

bush <- corpus_subset(data_corpus_inaugural, President == "Bush" & FirstName == "George W.") \%>\% 
tokens(remove_punct = TRUE) \%>\% 
dfm() \%>\% 
textstat_frequency() \%>\% 
as.data.frame() \%>\% 
select(feature, frequency)

kld <- kldivergence_shift(reagan, bush)
}
\seealso{
Other shifts: 
\code{\link{entropy_shift}()},
\code{\link{jsdivergence_shift}()},
\code{\link{proportion_shift}()},
\code{\link{weighted_avg_shift}()}
}
\concept{shifts}
