% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shifts.R
\name{proportion_shift}
\alias{proportion_shift}
\title{Proportion Shift}
\usage{
proportion_shift(type2freq_1, type2freq_2)
}
\arguments{
\item{type2freq_1}{A data.frame containing words and their frequencies.}

\item{type2freq_2}{A data.frame containing words and their frequencies.}
}
\value{
Returns a list object of class shift.
}
\description{
Shift object for calculating differences in proportions of types across two systems.
}
\examples{
Example to follow.
}
\seealso{
Other shifts: 
\code{\link{weighted_avg_shift}()}
}
\concept{shifts}
