% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shifterator-package.R
\docType{package}
\name{shifterator-package}
\alias{shifterator}
\alias{shifterator-package}
\title{shifterator: Functionality for Constructing Word Shift Graphs}
\description{
What the package does (one paragraph).
}
\author{
\strong{Maintainer}: First Last \email{first.last@example.com} (\href{https://orcid.org/YOUR-ORCID-ID}{ORCID})

}
\keyword{internal}
