% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{user_reviews}
\alias{user_reviews}
\title{User reviews of Animal Crossing.}
\format{
A data frame with 53940 rows and 10 variables:
\describe{
\item{grade}{Raw score (0-10) where higher score = better.}
\item{user_name}{User name of reviewer.}
\item{text}{Raw text of the review.}
\item{date}{Date review published.}
}
}
\source{
\url{https://github.com/rfordatascience/tidytuesday/blob/master/data/2020/2020-05-05/readme.md}
}
\usage{
user_reviews
}
\description{
A dataset containing user reviews of the game Animal Crossing: New Horizons.
}
\keyword{datasets}
