% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shifts.R
\name{proportion_shift}
\alias{proportion_shift}
\title{Proportion Shift}
\usage{
proportion_shift(type2freq_1, type2freq_2)
}
\arguments{
\item{type2freq_1}{A data.frame containing words and their frequencies.}

\item{type2freq_2}{A data.frame containing words and their frequencies.}
}
\value{
Returns a list object of class shift.
}
\description{
Shift object for calculating differences in proportions of types across two systems.
}
\details{
The easiest word shift graph that we can construct is a proportion
shift. If \eqn{p_i^{(1)}} is the relative frequency of word \emph{i} in the first
text, and \eqn{p_i^{(2)}} is its relative frequency in the second text, then
the proportion shift calculates their difference:

\strong{\eqn{\delta p_i = p_i^{(2)} - p_i^{(1)}}}

If the difference is positive (\eqn{\delta p_i > 0}), then the word is
relatively more common in the second text. If it is negative (\eqn{\delta p_i < 0}),
then it is relatively more common in the first text. We can rank words by
this difference and plot them as a word shift graph.
}
\examples{
#' library(shifterator)
library(quanteda)
library(quanteda.textstats)
library(dplyr)

reagan <- corpus_subset(data_corpus_inaugural, President == "Reagan") \%>\% 
  tokens(remove_punct = TRUE) \%>\% 
dfm() \%>\% 
textstat_frequency() \%>\% 
as.data.frame() \%>\% # to move from classes frequency, textstat, and data.frame to data.frame
select(feature, frequency) 

bush <- corpus_subset(data_corpus_inaugural, President == "Bush" & FirstName == "George W.") \%>\% 
tokens(remove_punct = TRUE) \%>\% 
dfm() \%>\% 
textstat_frequency() \%>\% 
as.data.frame() \%>\% 
select(feature, frequency)

prop <- proportion_shift(reagan, bush)
}
\seealso{
Other shifts: 
\code{\link{entropy_shift}()},
\code{\link{jsdivergence_shift}()},
\code{\link{kldivergence_shift}()},
\code{\link{weighted_avg_shift}()}
}
\concept{shifts}
