% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shifterator-package.R
\docType{package}
\name{shifterator-package}
\alias{shifterator}
\alias{shifterator-package}
\title{shifterator: Functionality for Constructing Word Shift Graphs}
\description{
The shifterator package provides functionality for constructing word shift graphs, horizontal bart charts that quantify which words contribute to a pairwise difference between two texts and how they contribute. By allowing you to look at changes in how words are used, word shifts help you to conduct analyses of sentiment, entropy, and divergence that are fundamentally more interpretable.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/pverspeelt/shifterator}
  \item Report bugs at \url{https://github.com/pverspeelt/shifterator/issues}
}

}
\author{
\strong{Maintainer}: Philippe Verspeelt \email{phiver4@gmail.com}

}
\keyword{internal}
